creacion de modelos y migrate de articulos lotes y pedimentos

+++ database/migrations/2017_10_05_044609_create_inventarios_table.php
@@ -19,7 +19,7 @@ class CreateInventariosTable extends Migration
             $table->bigInteger('PedimentoID');
             $table->bigInteger('LoteID');
             $table->bigInteger('ArticuloID');
-            $table->decimal('Cantidad');
+            $table->float('Cantidad');
             $table->timestamps();
         });
     }
